chore: Update .gitignore and add timeline_uk.js
This commit updates the .gitignore file to include the _site directory and adds a new file, timeline_uk.js, to the static directory. The timeline_uk.js file imports functions from timeline_main.js and timeline_uk_data.js and runs the timeline using the dateContainers data. This change is necessary to enhance the functionality of the England NHS Timeline.